Mark owned collections composite key properties use generated values even after the first two usages
Fixes #20932

The first time a CLR type is used as an owned collection it is not weak. Value generation is set correctly.
The second time, it becomes weak, and value generation is ste correction for both usages.
The first time, it is already weak, but the FK is marked as IsOwnersShip after value generation has been set.

The fix to run the value generation convention again when FK ownership changes.